add: tailwind to Abilities
