+++ src/assets/index.css
@@ -4,9 +4,11 @@
 
 @layer components {
 
-  .sidebar-icon {
-    @apply relative flex items-center;
-    /* ... */
+  .abilitycard {
+    @apply relative rounded-tl-sm rounded-tr-sm rounded-br-full rounded-bl-full bg-ability-card
+           p-4 flex flex-col items-center gap-4
+           border border-solid border-black shadow-abilitycard
+           hover:scale-110 transition-transform ease-in-out duration-150;;
   }
 
 }
