--- src/components/AbilityCard/_abilitycard_style.scss
@@ -1,40 +1,0 @@
-$img_size: 9rem;
-
-.abilitycard {
-    border-radius: 2px 2px 150px 150px;
-    border: 1px solid black;
-    background-color: gold;
-    padding: 1rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1em;
-    position: relative;
-    box-shadow: 0 5px 10px rgba( #000, .8 );
-    transition: all .2s ease;
-
-    img {
-        width: $img_size;
-        height: $img_size;
-        object-fit: contain;
-
-        -webkit-user-drag: none;
-        -khtml-user-drag: none;
-        -moz-user-drag: none;
-        -o-user-drag: none;
-        user-drag: none;
-    }
-
-    .abilitycard__description {
-        max-width: $img_size;
-        text-align: center;
-        margin-bottom: 1.2rem;
-    }
-}
-
-.abilitycard:hover {
-    transform: scale(1.1);
-
-    box-shadow: 0 2px 4px rgba( #000, .8 );
-    // box-shadow: rgba(255, 251, 10, 0.226) 0px 1px 1px 0px, rgba(251, 255, 14, 0.25) 0px 50px 100px -20px, rgba(255, 251, 10, 0.3) 0px 30px 60px -30px;
-}
