--- src/pages/Home/components/Abilities/_abilities_style.scss
@@ -1,15 +1,0 @@
-.abilities__container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-    box-shadow: inset 0 0 10px #424242;
-    padding: 30px;
-    gap: 2rem;
-}
-
-@media (max-width: 50rem) {
-    .abilities__container {
-        height: 70vh;
-        overflow-y: scroll;
-    }
-}
