--- src/pages/Home/components/AboutMe/_aboutme_style.scss
@@ -1,93 +1,0 @@
-.aboutme__container {
-    display: grid;
-    grid-auto-flow: row;
-    gap: 1em;
-}
-
-@media(min-width: 50em) {
-    .aboutme__container {
-        grid-template-columns: repeat(3, 1fr);
-    }
-}
-
-.life__chart {
-    justify-self: center;
-    width: clamp(200px, 50vw, 300px);
-    height: clamp(200px, 50vw, 300px);
-    align-self: end;
-
-    padding: 10px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 5px;
-    position: relative;
-}
-
-.student__half {
-    flex: 1;
-
-    height: 96%;
-    align-self: flex-end;
-
-    border-radius: 400px 0 0 400px;
-    // border-radius: 200px 0 0 200px;
-
-    background-color: #F7DD72;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-.student__half::before {
-    content: "Student";
-    font-size: clamp(0.7em, 4vw, 1em);
-    margin-bottom: 4%;
-}
-
-.programmer__half {
-    flex: 1;
-    height: 100%;
-    border-radius: 0 400px 400px 0;
-
-    background-color: #00F0B5;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-.programmer__half::before {
-    content: "Programmer";
-    font-size: clamp(0.7em, 4vw, 1em);
-    position: absolute;
-}
-
-.student__section,
-.programmer__section {
-    font-size: clamp(1.1em, 4vw, 1.2em);
-}
-
-.student__section {
-    ul {
-        list-style-position: inside;
-        margin-left: 1rem;
-    }
-}
-
-.programmer__section {
-    text-align: end;
-
-    .interests__list {
-        width: 210px;
-        float: right;
-
-        ul {
-            list-style-position: inside;
-            text-align: start;
-        }
-    }
-}
